backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
openpyxl>=3.1.0
python-dateutil>=2.8.2
openai>=1.0.0
rapidfuzz>=3.0.0
pandas>=1.5.0
